Use `fileURL` throughout rather than `filePath`

+++ src/files.js
@@ -4,7 +4,7 @@ import fs, { readFileSync } from 'fs'
 import path from 'path'
 
 import { t } from 'plottr_locales'
-import { actions, reducers, emptyFile, selectors } from 'pltr/v2'
+import { helpers, actions, reducers, emptyFile, selectors } from 'pltr/v2'
 
 import { closeDashboard } from './dashboard-events'
 import { store } from './app/store'
@@ -67,7 +67,8 @@ export const newFile = (
   }
   return uploadToFirebase(emailAddress, userId, file, fileName).then((response) => {
     const fileId = response.data.fileId
-    openFile(`plottr://${fileId}`, fileId, false)
+    const fileURL = helpers.file.fileIdFromPlottrProFile(fileId)
+    openFile(fileURL, false)
     closeDashboard()
     return fileId
   })
@@ -91,45 +92,41 @@ export const messageRenameFile = (fileId) => {
 
 // FIXME: we should get to a point where `whenClientIsReady` is
 // injected everywhere.
-export const saveFile = (filePath, file) => {
+export const saveFile = (fileURL, file) => {
   return whenClientIsReady(({ saveFile }) => {
-    return saveFile(filePath, file)
+    return saveFile(fileURL, file)
   })
 }
 
-export const editKnownFilePath = (oldFilePath, newFilePath) => {
-  ipcRenderer.send('edit-known-file-path', oldFilePath, newFilePath)
+export const editKnownFilePath = (oldFileURL, newFileURL) => {
+  ipcRenderer.send('edit-known-file-path', oldFileURL, newFileURL)
 }
 
 const win = getCurrentWindow()
 
 export const showSaveDialogSync = (options) => dialog.showSaveDialogSync(win, options)
 
-const escapeFileName = (fileName) => {
-  return escape(fileName.replace(/[/\\]/g, '-'))
-}
-
-export const offlineFilePathFromFileName = (filePath) => {
-  const fileName = escapeFileName(filePath)
-  return path.join(OFFLINE_FILE_FILES_PATH, fileName)
-}
-
-export const offlineFilePath = (file) => {
-  if (!file?.file?.fileName) {
+export const offlineFileURLFromFile = (file) => {
+  if (!file?.project?.fileURL) {
     return null
   }
 
-  if (file.file.fileName.startsWith(OFFLINE_FILE_FILES_PATH)) {
-    return file.file.fileName
+  const fileURL = file?.project?.fileURL
+  if (helpers.file.withoutProtocol(fileURL).startsWith(OFFLINE_FILE_FILES_PATH)) {
+    return fileURL
   }
 
-  return offlineFilePathFromFileName(file.file.fileName)
+  if (!helpers.file.urlPointsToPlottrCloud(fileURL)) {
+    return null
+  }
+  const fileId = helpers.file.fileIdFromPlottrProFile(fileURL)
+  return helpers.file.filePathToFileURL(path.join(OFFLINE_FILE_FILES_PATH, fileId))
 }
 
-export const renameFile = (filePath) => {
-  if (filePath.startsWith('plottr://')) {
+export const renameFile = (fileURL) => {
+  if (helpers.file.urlPointsToPlottrCloud(fileURL)) {
     const fileList = selectors.knownFilesSelector(store.getState().present)
-    const fileId = filePath.replace(/^plottr:\/\//, '')
+    const fileId = fileURL.replace(/^plottr:\/\//, '')
     if (!fileList.find(({ id }) => id === fileId)) {
       logger.error(`Coludn't find file with id: ${fileId} to rename`)
       return
@@ -140,15 +137,16 @@ export const renameFile = (filePath) => {
   const fileName = showSaveDialogSync({
     filters,
     title: t('Give this file a new name'),
-    defaultPath: filePath,
+    defaultPath: fileURL,
   })
   if (fileName) {
     try {
-      let newFilePath = fileName.includes('.pltr') ? fileName : `${fileName}.pltr`
-      editKnownFilePath(filePath, newFilePath)
-      const contents = JSON.parse(readFileSync(filePath, 'utf-8'))
-      saveFile(newFilePath, contents)
-      moveItemToTrash(filePath, true)
+      const newFilePath = fileName.includes('.pltr') ? fileName : `${fileName}.pltr`
+      const newFileURL = `device://${newFilePath}`
+      editKnownFilePath(fileURL, newFileURL)
+      const contents = JSON.parse(readFileSync(fileURL, 'utf-8'))
+      saveFile(newFileURL, contents)
+      moveItemToTrash(fileURL, true)
       store.dispatch(actions.applicationState.finishRenamingFile())
     } catch (error) {
       logger.error(error)
@@ -158,17 +156,17 @@ export const renameFile = (filePath) => {
   }
 }
 
-export const renameCloudBackupFile = (fileName, newName) => {
-  return fsPromises
-    .rename(offlineFilePathFromFileName(fileName), offlineFilePathFromFileName(newName))
-    .catch((error) => {
-      // Ignore errors renaming the backup file.
-      return true
-    })
-}
+export const deleteCloudBackupFile = (fileURL) => {
+  if (
+    !helpers.file.isDeviceFileURL(fileURL) ||
+    !helpers.file.withoutProtocol(fileURL).startsWith(OFFLINE_FILE_FILES_PATH)
+  ) {
+    return Promise.reject(
+      new Error(`Attempted to delete an offline file for non-offline file: ${fileURL}`)
+    )
+  }
 
-export const deleteCloudBackupFile = (fileName) => {
-  const filePath = offlineFilePathFromFileName(fileName)
+  const filePath = helpers.file.withoutProtocol(fileURL)
   return fsPromises.unlink(filePath).catch((error) => {
     // Ignore errors deleting the backup file.
     return true
